requirements.txt
streamlit
botorch
gpytorch
torch
pandas
numpy
matplotlib
gspread
google-auth